arm: apple: nvme: Add SART support and RTKit buffer management
The NVMe firmware in the macOS 13 beta blocks or crashes with u-boot's
current minimal RTKit implementation. It does not provide buffers for
the firmware's buffer requests. The ANS2 firmware included in macOS 11
and 12 tolerates this. The firmware included in the first macOS 13 beta
requires buffers for the crashlog and ioreport endpoints to function.

In the case of the NVMe the buffers are physical memory. Access to
physical memory is guarded by what Apple calls SART.
Import m1n1's SART driver (exclusively used for the NVMe controller).
Implement buffer management helpers for RTKit. These are generic since
other devices (none in u-boot so far) require different handling.

Signed-off-by: Janne Grunau <j@jannau.net>

Cover-letter:
Improve robustness of NVMe suuport for Apple silicon devices
Hej,

this series has assorted fixes to improve the robustness of the NVMe
support on Apple silicon devices. The main change which prompted this
series is "rm: apple: nvme: Add SART support and RTKit buffer
management". It fixes the RTKit driver required for the NVMe with the
system-wide firmware included in the first macOS 13 beta release.

The increased timeouts in RTKit are a defensive change against future
changes. None of the mailbox receive calls is expected to timeout so
waiting up to 1 second for a fatal error seems acceptable.

cheers Janne
END
Series-to: u-boot@lists.denx.de

Series-version: 2

Cover-changes: 2
 - update sart compatible strings to the SoC specified ones from the
   Linux devicetree bindings

Commit-changes: 2
 - The generation compatible strings "apple,sart2" "apple,sart3" got
   droppend from the bindings merged for Linux 5.19. Use the SoC
   specific ones instead.